Simplified DecompositionDecl handling.
Let VarDecl do the init stuff and only insert the BindingDecls as a
specialty.
